refactor Color constants

+++ trunk/simulations-java/simulations/glaciers/src/edu/colorado/phet/glaciers/module/basic/BasicModule.java
@@ -32,7 +32,7 @@ public class BasicModule extends PiccoloModule {
     
     private static final double VIEW_X_SCALE = 0.5;
     private static final double VIEW_Y_SCALE = 0.1;
-    private static final Point2D VIEW_ORIGIN = new Point2D.Double( 500, 100 );
+    private static final Point2D VIEW_ORIGIN = new Point2D.Double( 500, 0 );
     
     //----------------------------------------------------------------------------
     // Instance data
